Adds the live_reload_port option.
This allows the javascript to work with a non default value for the
livereload port (specified with the port option in the guard command).

For example, in config/environments/development.rb in rails:
  config.middleware.insert_after(ActionDispatch::Static, Rack::LiveReload, :live_reload_port => 22351)
works with the following in the Guardfile:
  guard 'livereload', port: '22351' do
    ...
  end

This allows multiple guard livereload processes to operate at the same time without
conflicts.